Enforce world borders on all maps
